S18 Fixing Category routes

+++ api/src/controllers/categoryController.js
@@ -3,12 +3,11 @@ const fetch = require("node-fetch");
 const { uuid } = require("uuidv4");
 const { Category } = require("../db");
 const { Op } = require("sequelize");
+const { checkUuid } = require("../helpers/utils");
 
-//get all categories
-async function getAllCategories(req, res, next) {
+async function getCategories(_req, res, next) {
   try {
     const categories = await Category.findAll();
-    console.log(categories);
     return await res.send(categories);
   } catch (error) {
     next(error);
@@ -18,13 +17,19 @@ async function getAllCategories(req, res, next) {
 //create a category
 async function createCategory(req, res, next) {
   try {
-    req.body.id = uuid(); //BODY
-
-    const { name, id } = req.body;
+    const { name } = req.body;
+    const find = await Category.findOne({
+      where: {
+        name,
+      },
+    });
+    if (find) {
+      return res.status().send("Category Already Exists");
+    }
     //console.log("atts: ", id, name);
     const category = await Category.create({
-      id: id,
-      name: name,
+      id: uuid(),
+      name,
     });
     console.log(category);
     return await res.send(category);
@@ -36,15 +41,23 @@ async function createCategory(req, res, next) {
 //delete a category
 async function deleteCategory(req, res, next) {
   try {
-    const { id } = req.body; //BODY
-    const category = await Category.destroy({
-      where: {
-        id: id,
-      },
-    });
-    const categories = await Category.findAll();
-    console.log(categories);
-    return await res.send(categories);
+    const { id, name } = req.body; //BODY
+    if (checkUuid(id)) {
+      const category = await Category.destroy({
+        where: {
+          id: id,
+        },
+      });
+      if (category === 0) {
+        return res.status(404).send("Id not found");
+      } else {
+        const categories = await Category.findAll();
+        console.log(categories);
+        return await res.send({ name: name, id: id });
+      }
+    } else {
+      return res.status(400).send("Invalid");
+    }
   } catch (error) {
     next(error);
   }
@@ -71,7 +84,7 @@ async function updateCategory(req, res, next) {
 }
 
 module.exports = {
-  getAllCategories,
+  getCategories,
   createCategory,
   deleteCategory,
   updateCategory,
